Move ViewGroup to its own module and move object chain implementations
